Finding the natural start and end of local maximum, I now need to do merging
